Workspace_Project_Version3/CamLight_Package/cam_status_mqtt/requirements.txt
paho-mqtt
pymongo
pytz